Gehennom changes (round two) - Cerberus guards the entrance to Gehennom. Another change to the Valley of the Dead - Cerberus now guards the stairs leading down to Gehennom. Cerberus was actually defined in src/monst.c and has been for a long time, but was never used and was subsequently commented out. This is not the default Cerberus template - he is much more of a threat here. Charon has also been taken out of storage as it were - his role is to simply guard the gate to Cerberus' lair, and to warn the hero of what lies ahead. The details:
* Cerberus guards the stairs to Gehennom. He has three powerful bite
attacks (as he has three heads) plus a single fire-based breath attack
(think of Cerberus as the lord of the hellhounds).
* Cerberus' attacks may not seem too dangerous, but coupled with the
fact that he is very fast, has high magic resistance and a low AC along
with several resistances, he is a force to be reckoned with.
* Charon is unchanged from the original template for the most part - he
spawns peaceful and will give useful information to the player if they
chat with him.
* The Valley of the Dead has had a slight redesign to accommodate
Cerberus' lair, which is located north of Moloch's temple. There is a
moat and drawbridge - the moat is the remnants of the river Styx. The
drawbridge acts as the unoffical Gates of Hell.
* This particular drawbridge cannot be destroyed - if the player does
not have a magical means of opening the drawbridge, they will not be
able to descend into Gehennom via the stairs. There are two guaranteed
wands of opening in the valley which the player may want to obtain.

I purposely did not incorporate anything from the Lethe patch for this,
as amnesia is already disdained by many players. No need to add even
more sources to the game. Some possible TO-DO's - allow Charon to open
the drawbridge for a considerable fee, and/or don't allow level
teleportation below the Valley unless the drawbridge has been opened.