Bypassing Cerberus. If Cerberus is guarding the entrance to Gehennom, it doesn't make a lot of sense that our hero could potentially bypass him by simply levelporting out of the Valley of the Dead and into Gehennom. Or say the player is able to get around Cerberus and wants to use a cursed potion of gain level to get past him on the way back up - that doesn't make sense either. What this commit does:
* Level teleporting no longer works while in the valley. The player can
levelport to the valley, but if they want to continue down into
Gehennom, they'll have to to use the stairs
* Cursed potion of gain level - drinking one on the level right below
the valley will put the player at the downstairs in the valley
(Cerberus' lair). Trying to drink one while in the valley won't work,
the player has to make their way to the upstairs
* Cerberus has teleport control - zapping him with a wand of teleport
just repositions him next to the player

Branchporting wasn't touched - the player can't branchport to a level
they've never been to before (making your way down), and the player
can't branchport with the Amulet of Yendor (ascension run back up).

There are ways to get past and/or defeat Cerberus, but our hero will
need to use their wits and be smart about it, plan things out. No
freebies getting to hell.