Some angels and priests on astral level have 'purple rain' insurance. A popular tactic some playrs use on astral level is called 'purple rain', in which the player has several cursed scrolls of genocide, several wands of teleportation with at least one charge left, and a means to easily tame multiple monsters at once. They reverse genocide purple worms, tame them, then break a wand of teleportation to scatter them about the level. They then act like vacuum cleaners, eating up as many angels and priests and other smaller monsters as they can before they succumb. Because dragons can also engulf in EvilHack, they've started to replace purple worms as the go-to monster to use for obvious reasons.
The denizens of the astral plane have had enough of being eaten alive,
and have taken measures to combat it. Priests on astral level have a
50% chance of spawning with a worn ring of slow digestion, angels have a
33% chance of the same. Being swallowed and regurgitated still does
damage, and the tactic described above is still a viable one, but it is
no longer a wholesale slaughter of almost every monster on the astral
plane without the player barely having to attack anything.