Gehennom changes (round four) - Vlad's tower. This is a complete revamp of Vlad's tower. The details:
* All three tower levels have been remade
* An extra fourth level has been created - a huge cavern that is
directly below the bottom tower level, where a large colony of 'vampire
bats' roost
* The stairs to access Vlad's tower in Gehennom go down into the new
cavern level instead of up as before
* The new cavern level is different than any other level found in the
game - the primary terrain is air, and gravity is still in effect,
meaning it is impossible to traverse this level unless the player can
levitate or fly by whatever means
* Because gravity is in effect and there is no floor for most of the
cavern level, anything dropped while in midair is lost... objects,
corpses, anything. Forever.
* There are many more monsters to have to fight to reach Vlad (mostly
vampires and other undead) - Vlad has grown tired of being defeated
constantly and has upgraded his defense posture
* The guaranteed items found in Vlad's tower are the same as before,
with a few extra new items added for good measure

The most time-consuming part of this revamp was making it so if the
player or monsters couldn't fly or float, they would fall to their
deaths in the new cavern level. Even more difficult was objects being
dropped or thrown while in the air. I plan on going back over all this
again (being life-saved when falling to your death comes to mind) but I
think most scenarios have been covered. The new cavern level is
especially dangerous because as it stands right now, if the Amulet of
Yendor is dropped in midair here, or if any of the invocation items are
dropped here and the invocation ritual hasn't been performed yet, the
game becomes unwinnable. Deciding on whether to let that stand or if a
special provision should be made for those items.

Overall, the intent was to make Vlad's tower more fun, exciting, and
dangerous. Hopefully that intent is met.