New monster: weredemon. Adding a new werecreature that can be found in the mid to late game, the weredemon. A hybrid between a demon and a hell hound, a weredemon is significantly tougher than a werewolf and can be a serious threat. In animal form, they appear as hell hounds and share its fire-breathing attack. In demon form they claw and bite, and their bite in this form can still confer lycanthropy. Because they are flagged MH_DEMON and MH_WERE, they can summon both hell hounds and other demons when in demon form. Unlike their humanized counterparts, weredemons do not leave corpses upon death.
I've sprinkled a few of these in the valley, Vlad's tower, and a couple
of the demon lord lairs. They can spawn outside of Gehennom, so it would
be possible to encounter these right around the time the player reaches
the castle.

I've made their demon form black in color so they stand out (reverted
pit fiend back to its original red color). Also a slight tweak to
wererat frequency in this commit.