Valkyrie quest artifact changes.
The Orb of Fate has been replaced with Gjallar, the horn used by the
norse god Heimdallr to warn the gods of Ragnarok. The details:

* Gjallar has all of the same special effects that the Orb of Fate had,
other than it is now a tooled horn rather than a crystal ball
* Gjallar is loud - when applied, it can be heard everywhere on the
current level it was used on. Because of this, the 'scare distance' from
the player applying Gjallar is of a greater distance than just a regular
tooled horn
* There is a chance that monsters adjacent to the player applying
Gjallar will become stunned for a short time
* If the player applying Gjallar is not a Valkyrie, there's a chance
they can make themselves deaf for a brief period