Changes to covetous monster behavior.
From SporkHack (commits 82e5b24, b83b4cf and c8c283b). Covetous monsters
will pursue the player using slightly random jumps. In wide-open spaces,
this gives the player a chance to attack with a ranged weapon/spell (as
well as the covetous monster), but it can also make it more difficult to
chase a covetous monster down. Teleporting to the upstairs is still part
of their strategy.