Chance the Amulet of Yendor will teleport away if dropped.
From UnNetHack (commit df1cb87), slightly modified - if the Amulet of
Yendor is dropped, either by the player or another monster, there is a
chance it will teleport away to another location on the same level. That
chance is determined by its beautitude (66% chance if cursed, 25% chance
if uncursed, 6% chance if blessed). Beware when dropping the amulet to
save on spell power when casting.