Major amnesia revamp.
From NetHack official build, version 3.7.0 (commit 04c59ff), modified -
instead of forgetting maps and objects, amnesia will cause the player to
forget skills. From the original commit:

'Forgetting skills allows the player to optionally go down another
skill path, if they trained the wrong weapon in the early game.

Amnesia still forgets spells.

As a replacement for the deja vu messages when entering a forgotten
level, those messages will now indicate a ghost with your own name
existing on the level, given only when the level is entered for
the first time.'

Changes from the original commit: traps will still be forgotten for that
level. AD_DRIN attacks will also cause skills to be drained (how many
depends on your luck), a chance for a certain number of spells to be
forgotten, and all traps on the current level along with the usual INT
drain.

This change to amnesia in vanilla NetHack made mind flayers a lot less
dangerous, since a unicorn horn can be used to restore lost attribute
points and nothing is forgotten at all. Between this spin on the amnesia
revamp and unicorn horns unable to cure attribute loss, this still keeps
mind flayers as a threat to be taken seriously.