New monsters: mind flayer larva and alhoon.
Two new monsters to add some flavor to the mind flayer species - mind
flayer larva and alhoon. Mind flayer larva are just that, the young form
of a mind flayer before they take over a host and metamorph into its
adult form. Imagine a baby squid or cuttlefish. Details on these:

* Mind flayer larvae are tiny, relatively weak and live in shallow pools
of water
* They do not generate randomly, and are only found in a special new zoo
known as a mind flayer nursery. These nurseries won't start to appear
until dungeon level 20 or later, and are rare. They are tended by an
adult mind flayer
* Mind flayer larvae appear as S_EEL and are the same color as mind
flayers
* Even though they are weak and easy to kill, mind flayer larvae are
slightly faster than average, and their attack is to attach themselves
to the players face and burrow into their brain
* This attack falls under AD_WRAP, and is very much like an eel's wrap
attack - the player will have exactly one turn to decide how to proceed
if a mind flayer larva is able to attach itself to the player's face.
After that one turn is up, the larva burrows into the player's brain,
taking over their psyche and killing them. If bones are left, they will
arise as an adult mind flayer

Alhoon are the product of rogue mind flayers attempting to gain ultimate
power as a lich. They still retain their form but are a sickly gray in
color. Alhoon are basically a cross between an arch-lich and a master
mind flayer - not as many tentacle attacks, but they can employ the
lich frost touch attack and can cast spells. They are flagged as
MH_UNDEAD and enjoy all of the benefits and weaknesses that come with
being undead. They are very rare, and typically will only be found in
Gehennom.

Will probably tweak a bit of this soon (illithid player race shouldn't
be affected by a mind flayer larva for instance). More to follow.