Fix: 'You slip the leash around your Fiesty's gnome mummy'.
Using a leash on a named pet now returns just their name instead of
'your <petname>'. The name of a bones monster other than a ghost is now
'<playername> the <monster>', seems more fitting. May revisit this
again, I feel this could be done up a little better, but this works and
it doesn't break saves.