New branch: The Ice Queen's Realm.
EvilHack's first ever new branch - The Ice Queen's Realm. The story -
the player finds a magic portal that transports them to a faraway land
that is trapped in an eternal winter. This land is filled with deadly
creatures never seen before. The player must find their way through this
land and find the source of this affliction upon the land, and destroy
it if at all possible.

Last year I promised my daughter (aged seven at the time) that I would
add her and our dogs to EvilHack. She loves the movie 'Frozen' likes to
pretend she has Elsa's ice magic. So that's how I worked her into this
branch - Kathryn the Ice Queen is the boss monster at the end of this
branch that the player must defeat in order to complete this quest. Our
german shepherd dogs (Koa and Ozzy) are her guardians. The player
doesn't actually kill the Ice Queen so much as defeats her, but that
isn't exactly the very end of the quest. Details:

* New monsters (snow golem, saber-toothed tiger, revenant, ice nymph,
frost salamander) along with a small handful of unique monsters
* The new monsters made for this branch will only ever spawn in this
branch and nowhere else. Only other way you would come across them is if
another monster polymorphed into one of these
* The Ice Queen's Realm is so cold, if the player does not have 100%
cold resistance, they will take a small amount of cold damage each time
they move (the player's amount of cold resistance determines the amount
of damage inflicted)
* New terrain object: dead trees (from UnNetHack, modified)
* The Ice Queen's spells have been geared towards cold-based attacks (in
doing this, fixed some minor bugs with monster spellcasting and other
types of monsters)
* The portal to the Ice Queen's Realm will spawn on the level right
before Medusa's level
* A new type of 'novel' was created that the player can find right as
they enter the branch - a journal left behind by another adventurer, the
player can gain insights and clues as to what they can expect if they
read it, and it can help them find their way to the Ice Queen
* The prize - a caged pegasus that the Ice Queen has captured. A unique
animal that will only ever be found here. Can be tamed and ridden if
freed.

This branch is purely optional. The player can skip this entirely if
they so choose, and will not have any effect on the outcome of the rest
of the game. This side branch is challenging, and the player should come
well-prepared if they wish to try it.

Along with this commit - entering Gehennom, if the player does not have
full 100% fire resistance, they will take a small amount of fire damage
just as they would take cold damage in the Ice Queen's Realm.

I'm sure this new branch will have issues to address and things to
tweak. I hope it's an enjoyable side quest for anyone attempting it.