Autounlock doors and containers.
Brought over from NetHack 3.7, commits 9b0d78d and bfa0f07 (modified).
That in turn was pulled from xNetHack, which had pulled it over from
UnNetHack. Heh.

If the player has the right kind of tool in open inventory, if they go
to open a locked door or container, the game will automatically ask if
they want to unlock it and then do so. There's some special handling
here for iron safes and stethoscopes.