Some tweaks to player monster container types, object property odds.
Player monsters shouldn't be viewed as the guaranteed endless source of
bags of holding. High level player monsters will now only spawn with
them a quarter of the time instead of all the time.

Adjusted the odds of weapons/armor spawning with an object property
again - odds are now 1 in 150 instead of 1 in 250, and for weapons, the
odds slightly get better the deeper into the dungeon the player goes.
Tweaks odds in other areas as well (player monster and S_ANGEL weapons
at spawn, and altar sacrifice gift if it's not an artifact).