Fix: poisoned weapons other than projectiles were not flagged correctly.
Awhile back, I made it so any slashing/piercing weapon could be poisoned
using a potion of sickness. I recently noticed that attacking with such
a weapon treated this way wasn't behaving as if it were poisoned. Where
the flag was set (ispoisoned = TRUE) was still only applied to
projectiles. This commit fixes that.