Fix: a couple issues with monsters casting spells against other monsters.
Monster casting CLC_BLIND_YOU against another monster was showing the
player the blinding feedback even if the player could not see the
recepient monster.

Monster summoning an elemental or insects against another monster would
be created around the player instead, and whatever was summoned would
still be hostile to the player. This is a much harder nut to crack, so
until a better solution is found, we'll reroute these spells to
something else if the casting monster is tame.