Greatly reduce the chance of a kick missing if the player is skilled in martial arts.
Since automatic kicking is a thing for monks at master level and higher,
this needed to be addressed as a missed kick means the monster will
'evade' and move to a new position. Instead of the typical 1 in 4 odds
of the target evading, for martial artists it's only a 1 in 50 chance of
the target evading.