Adjust loot in special room for the Mines' End 'sewers' level.
Based on player feedback - sweeten the pot a lil bit so the 'juice is
worth the squeeze' getting into the special hidden room at the Mines'
End sewers. Also added an extra teleportation trap so the player isn't
running all over the level trying to teleport in without teleport
control.