Fix: handling of monster riders if polymorphed and still riding their steed.
This commit came about via report from another player that had zapped a
wand of polymorph at a monster riding a steed. I believe it was a goblin
outrider - the player had zapped the rider, which turned into a full
grown green dragon, yet the dragon was still riding its wolf. Hah.

What happens now - if the new polyform is MZ_LARGE or larger, and if
it's not humanoid in shape, it will fall off its steed.

Also in this commit, a bit of correction for when a rider is knokced
off its steed via being clobbered.