Fix: segfault when ball & chain rusts completely away under certain circumstances.
From xNetHack commit d820b66, which addresed potions of acid exploding
in water. The backtrace of the player's crash that prompted that fix was
almost identical to the backtraces I have when a segfault occurs when a
ball & chain rusts away. After several fuzz sessions, I'm not seeing
those crashes anymore.