Tweak damage bonus for using two-handed weapons, other code clean-up.
In a few places, there were expressions similar to this: tmp *= 1.5;
meaning take whatever value tmp is, and multiply it by 1.5 - not the
best way to do this. This commit corrects many of these, and will
resolve some warnings on the windows binary compile side.

Tweaked the damage bonus slightly based on strength - values are the
same, but one cut-off was adjusted down to 18/50 instead of 18/75 (this
is in the player's favor). And now, instead of just getting an extra
damage bonus for wielding a weapon with two hands, its truly wielding a
two-handed weapon, meaning giants one-handing a two-handed weapon will
enjoy the slight damage bonus (two-handed weapons have a fair amount of
mass).