Fix: dying and then being life-saved via roasting/freezing to death.
If the player was wearing an amulet of life-saving and they died from
either roasting from the heat in Gehennom or freezing from the cold in
the Ice Queen branch, their life would be saved, but then they'd lose
all hit points immediately right after and die again from the same
thing. In wizmode you'd be stuck in an infinite loop unless you chose to
die. This commit fixes that.