Adjust one type of DSM for player monsters.
Reason - there are some outstanding issues with duplicate light sources
crashing the game. The main culprit is when a monster is wearing gold
dragon scales/scale mail, and that happens most often when a player
monster spawns with it as part of their kit. Changing it for now until a
resolution is found for the duplicate light source crashes.