Fix: spcial artifacts should not be wished for.
Both Lifestealer (Vlad) and the Magic 8-Ball (Oracle) are special
artifacts that are always guaranteed to appear, and do not count towards
the number of existing artifacts in the game when it comes to wishing
for artifacts. Because they are special and are meant to be found with
their owner, make them non-wishable (in wizmode they can still be
wished for). This also resolves a rn2(0) bug if you tried to wish for
either of these before any other artifact existed.

The Sword of Bheleu will eventually make its way onto this list once the
new side branch that features it as the prize is created.