Another Juiblex tweak.
Juiblex is basically a super-evil steaming pile of acidic goo, so these
new additions make sense... to me at least :)

Juiblex now has a passive acid attack identical to that of a
spotted/oche jelly, which can corrode objects that are susceptible to
acid damage. Juiblex will also explode when killed (from the outside,
not engulfed), showering anything adjacent to it with acid.

Enjoy.