Fix: monsters switching back and forth between melee and ranged.
Before this commit was commit a0cae1e, which was dubbed a partial fix
for monsters stuck in a loop, switching back and forth between their
melee weapon and launcher. I think this commit is the actual fix.
Testing this produced the desired behavior, that being a monster attacks
you at range and won't advance upon your position and switch to melee
until it has exhausted its ranged options. We'll see how this goes.