Fix: bug with Medusa's gaze being reflected back and the player being blind.
Discovered that if the player had reflection, was blind and had no
telepathy, Medusa's reflected gaze would never have a chance to stone
her because the player couldn't see her. That's corrected in this
commit.

Also in this commit - Medusa can fly again (turns out according to Greek
mythology, she did have wings... go figure), she spawns with a random
number of arrows instead of a fixed number, and if the distance is too
great betwwen her and the player and the player has reflection, her
reflected gaze is too far away for her to notice. Was told she was still
a pushover, as one could just stand off several spaces away and wait for
the 5% chance of her reflected gaze to stone her. Alrighty then. That
can still happen, but you'll have to get close (almost but not quite
melee range) to do it.