Adjust material for various artifacts.
A fair amount of artifacts that exist are no good for elves or orcs
because of their aversion to iron (elves) and mithril (orcs). I've
adjusted the material for several to metal, including a couple of quest
artifacts - especially the Mitre of Holiness, as elves can be priests
and having your quest artifact made of iron would be a bad thing.

Also in this commit - I've opened up the heavy war hammer as a weapon
that can spawn randomly, and is no longer resigned to just the artifacts
Ogresmasher and Mjollnir. Because of the edged weapon restriction for
priests, this made the mace and the quarterstaff their highest damage
dealing weapon choices, and that's fairly weak versus a lot of other
weapons out there. A heavy war hammer is comparable to a long sword in
terms of damage output.