Redo what happens when giants bring boulders into Sokoban to drop/throw.
After talking to a few players about this and thinking on it some more,
decided to allow giants to bring in and drop/throw boulders all they
want. It'll just incur a luck penalty just as any other form of
'cheating' does while in Sokoban. But because it's easy for giants to
manipulate boulders (unlike other races), the luck penalty is twice that
of the usual Sokoban luck penalties.