Fix: certain metallic shields did not incur a substantial casting penalty.
Per vanilla NetHack code, if a shield weighs more than a small shield
(30 aum), the player wearing it will take a significant hit to their
chances to successfully cast spells. This allowed players to get away
with wearing a mithril shield of reflection (16 aum) and being affected
the same as if they were wearing a small shield. That's pretty
overpowered, and last I checked, anything metal was supposed to hinder
spellcasting anyways. So lets factor in shields that are metallic. Sorry
folks :)