Steel.
This is something that's kind of irked me for as long as I've been
playing NetHack - METAL has always been considered the catch-all
material for any metallic object that wasn't already specifically
defined. 'Metal' is a generic term however, and has led to various
interpretations as to what it actually is. The wiki entry for metal is
just 'a material used for metallic objects that are not iron, copper,
platinum, silver, or gold'. In doc/fixes34.0, line 414, metal is
referred to as stainless steel. Some players consider it to be more like
aluminum, or something else that's never really been defined.

With this commit, I've decided to equate the material METAL with steel,
and objects will be described as such if they spawn as being made of
METAL and that is not its base material. It will still keep its
naturally non-corroding/non-rusting properties (read: stainless). It's
density was bumped up a bit, but is still slightly less than iron. I
know that in real life, stainless steel is slightly more dense than iron
and is not as well suited for weapons as high-carbon steel would be, but
I'm allowed some creative license, and it's an extra reason to prefer
steel over iron weapons and armor.

The chances for various objects spawning as steel have been adjusted.
All weapons made of steel get a +1 damage bonus. Steel armor, although
inherently fixed, is subject to the monster spell 'destroy armor' and
can deteriorate (should have been this way from the start).