Fix: tame chameleons able to shapeshift after being 'killed' after a save/reload.
Over a year ago I had made it so that if the player killed a chamelon or
other shapesifter, instead of dying it would revert to its original form
and stay in its original form until killed. This was to give a little
bit more of a challenge to the player, as polymorphed players behave the
same way (if a player is polymorphed, their poly'd form 'dies' and they
revert back to their original form and live).

It's now been discovered that if the player has a tame chameleon, and
its poly'd form is killed, it stays in its base form (as intended), but
if the player saves/reloads their game, the chameleon is free to change
shape all over again. This has already been exploited, giving the player
in essence a near-unkillable pet so long as they monitor their pet
chameleon and save/reload when they notice it being reverted to its base
form. This commit corrects this oversight.

This same behavior occurs with non-tame shapeshifters, and it's present
in vanilla NetHack and at the very least any variant that is based off
of 3.6/3.7, so if this isn't the devteam's intent, this warrants a bug
report to them as well.