Fix: monster spell 'vulnerability' not working vs other monsters.
Mobileuser discovered that if a spellcasting monster cast its elemental
vulnerability spell at another monster, the player was affected instead
of the targeted monster. This commit corrects that behavior.

Also in this commit - needed to be able to make sure this fix was
working as intended, so now using a stethoscope or a wand of probing on
a monster will reveal any M$ flags that are present.