Fix: non-flying steeds falling to their death in Vlad's cavern.
If the player is flying and they mount a steed, the steed is then able
to fly also if it's not normally able to fly. Riding your non-flying
steed into the open air in Vlad's cavern would cause the steed to
plummet to its death. This commit fixes that. Thanks microlance for
discovering and reporting this.