Fix: monster rider/steed tweaks and fixes.
Good portion of this is from SpliceHack commit 7febec7 - it addresses
monster rider and/or their steeds trying to jump down a trap door to
esacpe. This also addresses monsters trying to mount a steed that is
already dead. Better handling for noncorporeal/unsolid monsters and
preventing those from trying to mount a steed.