Fix: heavy iron ball and chain are always rustproof.
There are currently six separate and distinct crashes that can occur if
a heavy iron ball or its chain completely rust away and conditions are
right. At some point I want to address these, but I haven't been able to
solve these yet and tournament time is fast approaching. With this
fix, I can run the fuzzer for millions of turns instead of a few
tens of thousands before experiencing an issue. This is the proverbial
band-aid for the time being.