Fix: "warning: address of array 'killer.name' will always evaluate to 'true'"
Thanks riker for finding this one - this commit silences a warning with
the whereis code. Before, 'ascended=' was always 1 no matter what. With
this minor fix, it actually changes between different values like it was
intended to do.

Also in this commit, a fair amount of code formatting cleanup with
whereis and the wish tracker sections of code.