Fix: monsters killed while standing on shallow water or sewage weren't leaving a corpse.
I probably stared at these sections of code for a lot longer than i
should have, it just never jumped out at me. Then riker looks at this
for five minutes and figures it out :) This commit now makes it so that
if a monster is killed (by the player or another monster) while standing
in shallow water or sewage, they will leave a corpse if they were meant
to do so.