Fix: warning C4244: '=': conversion from 'int' to 'float', possible loss of data.
In src/worn.c concerning adjustments to a monsters AC based on its
level. Was using a float where an int would suffice. This warning pops
up during a travis-ci build for the windows binary. There's a similar
warning in src/potion.c for partial resistances, but simply changing
float to int there isn't the answer (did some formatting/cleanup in that
section).