Fix: The gnome lord's silver pick-axe shatters from the force of your blow!
I have noticed for quite awhile now that the feedback received for
breaking a monsters wielded weapon was not entirely appropriate
depending on the material of said weapon. Some metals just don't
shatter, like silver or gold. They would bend instead, but I also don't
feel like going down that road (yet) for having a weapon damaged to the
point that it was bent. So for ceratin metals, the feedback will just be
'breaks' as that can actually happen in whatever form or fashion.

Mithril and crystal weapons have been made break-proof, as each material
is strong enough to withstand being broken during use in combat.