Fix: beheading a troll that was wearing an amulet of life saving shouldn't cancel it.
If a zombie or troll is beheaded, they are flagged as cancelled and will
not revive (no head, no reviving). If a troll was decapitated whilst
wearing an amulet of life saving, it was still flagged as cancelled, and
killing it again by normal means would still prevent it from reviving.
This commit sets a condition so they aren't flagged as cancelled if
beheaded and wearing an amulet of life saving. The same condition is not
set for zombies, as undead cannot be life-saved.