All spell staves are now properly magical.
Had to add a new bit to the WEAPON array in objects.c to flag weapons as
either magical or non-magical, just as many other objects already are.
So now, all of the spell staves are flagged as being magical, as they
should be. All other weapons are flagged as non-magical.