Disclose achievements.
From NetHack 3.7.0, modified (git commit a453215). Previously, in-game
achievements were only expressed as a bitfield in the xlogfile and not
human-readable. Now, achievements are also shown in the dumplog that
anyone can easily read and see what achievements were completed. The
player can also use the #conduct command in-game to view which
achievements they have earned so far.