Snow golems can throw snowballs.
Snow golems now have the ability to throw snowballs at their target. The
damage dealt is physical and not cold-based, and snowballs have the
capability to blind their target if said target has eyes. This attack is
an extension of AT_SPIT, but the feedback received is that of something
thrown.

Nothing game-changing or awe-inspiring here, just wanted to do this for
some added flavor and laughs when exploring the ice queen branch.