New player role: Infidel (by Tomsod).
Introducing a new role to EvilHack - the Infidel, created by Tomsod.
From Tomsod's user page on the NetHackWiki: 'Infidels are a role that
specializes in offensive magic, with some similarities to Priests, and
to Flame Mages of SLASH'EM. They can be humans or orcs; like Priests,
they have a random pantheon, but Infidels always follow Moloch. Infidel
role offers a slightly unusual game, as they start with the Amulet of
Yendor and are initially tasked with returning it to Moloch's Sanctum.'
More detailed information on their work on the Infidel role can be found
at https://nethackwiki.com/wiki/User:Tomsod/YANIs_and_patches/Infidel.

For the most part, this is a straight copy of Tomsod's latest patch
(version 1.3 as of this commit). There were a few bugs I noticed and
fixed along the way - most were code formatting issues that would ause
the compiler to complain, a couple were either issues that prevented
compilation to finish, or were crash bugs during normal play. In this
commit are also a few changes to make this role fit in with how EvilHack
works (e.g. Infidels won't start with a magic marker). There are more
changes and edits I'd like to do soon - illithids would be a great fit
for this role. Some other quirks and wording I'd like to change as well,
but I want to confer with Tomsod first before I do anything else.

This should be a fun and interesting role, and challenging as well. Will
play test this role a fair amount before the new version of EvilHack
goes live.