Various fixes and tweaks to the Infidel role.
There's a fair amount to this commit, but here are the highlights:

* Infidels start with two scrolls of charging in lieu of not having a
starting magic marker
* During initial play-testing, it became readily apparent that both the
Amulet of Yendor and the Idol of Moloch (once imbued) prevented the
player from teleporting in any way, and also causing the power drain on
spellcasting. I'm not sure if this was the role author's intent or not,
but I do know that players won't tolerate either behavior for the entire
game. The change is this - Infidel's are not affected by the Amulet's
teleportation suppression nor it's spell power drain; Moloch shields its
chosen one from those effects. But once the Amulet is sacrificed and the
Idol of Moloch assumes its power, those effects are amplified and affect
the player just as the Amulet would affect any other role.
* The Idol of Moloch cannot be destroy via dropping it in Vlad's cavern
or whacking a black dragon with it. If this item were destroyed, this
would make the game unwinnable for an Infidel
* If monsters come into possession of the Idol of Moloch, they won't
activate like they would a regular figurine (I watched the Paladin do
exactly this during play-testing. Oops.)
* Took some creative license, changed the crowning title for Infidel
from 'a demon of Moloch' to 'the Archfiend of Moloch'. Per Dungeons &
Dragons lore, an Archfiend is any of the several demon/devil overlords.
Totally appropriate here to put our freshly-crowned Infidel on the same
footing.
* Various minor tweaks and bug fixes (thanks Tomsod and amateurhour for
the feedback).

Co-authored-by: Tomsod <Tomsod@users.noreply.github.com>