Realizing the Infidel role author's intent.
Or at least trying to. Here are the changes:

* Reverted the changes to the Amulet of Yendor not suppressing
teleportation for Infidels if carried. If you want to teleport
somewhere, you'll need to drop the Amulet.
* Spell power drain is now reversed - the Amulet will drain spell power
while it's in the players inventory. Once it is sacrificed to Moloch and
the Idol of Moloch is imbued with its power, Moloch is able to suppress
that aspect of its power, and allows the Infidel to realize their full
potential.
* Small tweak the the Amulet of Yendor's base material, changed it from
mithril to crystal (gemstone). Reason being, orcish Infidels are a
thing, and they start the game wearing the Amulet. Orcs and mithril
don't mix.

The spell power draining switch may seem controversial, since the
Infidel will be able to cast spells with impunity during the ascension
run, unlike every other role. I think this is fair, since the Infidel
has to tolerate having their spell power drained for the other 95% of
the game versus everyone else's 5%.

Looking forward to the feedback, and hopefully this is more inline with
Tomsod's intent.