Fix: a few more incorrect alignment type reporting, livelogging (Infidel role).
Found a few more cases where an Infidel's alignment wasn't recorded
properly (xlogfile, whereis and livelogging). Also discovered during
play testing that on turn one, the livelog event 'acquired the Amulet of
Yendor' was triggered as soon as a game was started as an Infidel -
that's been suppressed, and a proper livelog message has been created
for when an Infidel imbues their quest artifact.