Effects of praying/sacrificing as an Infidel.
Moloch doesn't do shiny effects or light soothing colors, or even hand
out blessings (although he will still aid his own follower in their time
of need). The changes in this commit:

* Feedback for various effects due to praying are changed (e.g. instead
of being surrounded by a shimmering glow, for Infidels it's 'an ominous
crimson glow', etc).
* Moloch doesn't do blessings. Successfully praying on an unaligned
altar with potions of water on it as an Infidel produces cursed water
instead of blessed.
* The Infidel player can still have troubles fixed just as any other
role, just that the feedback is different. Cursed objects can still be
uncursed as well (except for armor, but this was already handled in the
initial patch).
* Instead of a wielded weapon being blessed, it's cursed. In vanilla
NetHack this would have zero benefit, but in EvilHack there are some
advantages to having a cursed weapon.
* Altar sacrificing - should the Infidel receive a regular piece of
gear, care has been taken to make sure that object is never made of
silver.