Crowned Infidels (demonic form) can never handle silver objects.
In vanilla NetHack, a player in creature form that hates silver could
never wield anything made of silver, even if wearing gloves. That fit in
nicely with the Infidel role authors intent, that when they were crowned
and changed into a demon, they would be under the same restrictions. In
EvilHack, and also in a couple other variants, if the player were of a
race or in creature form of something that hated a certain material,
they could wield an object made of that hated material if they were
wearing gloves. That won't work for how a crowned Infidel was
envisioned. This commit addresses this, and also ensures they can still
use the Bell of Opening to perform the invocation ritual.